fix(spells): Moved misplaced spell form
